# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pro POS API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /auth/owner/login:
    post:
      summary: Owner login without requiring POS activation
      responses:
        '200': { description: Owner session created }
        '401': { description: Invalid credentials }
  /auth/employee/login:
    post:
      summary: Employee PIN login bound to an ACTIVE POS device
      responses:
        '200': { description: Employee session created }
        '401': { description: Device or credentials invalid }
...